backend/requirements.txt
# SafeFlow Global — Backend Dependencies

# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9

# Database
asyncpg>=0.31.0
sqlalchemy[asyncio]==2.0.35
alembic==1.13.2

# Redis & Task Queue
redis==5.1.0
celery[redis]==5.4.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2

# AI
google-generativeai==0.8.2

# Notifications
twilio==9.2.0

# Validation & Utils
pydantic==2.9.0
pydantic-settings==2.5.0
python-dotenv==1.0.1
httpx==0.27.2

# CORS
starlette==0.38.6
